SSL is required (403 Forbidden)
The Twitter API now requires SSL. Add ( ssl => 1 ) to the options passed to new
to enable it.  For backwards compatibility, SSL is disabled by default in this
version. Passing the ssl option to new will disable this warning. If you are
using a Twitter API compatbile service that does not support SSL, add
( ssl => 0 ) to disable this warning and preserve non-SSL connections in future
upgrades.